requirements.txt
# المكتبات الأساسية المطلوبة
python-telegram-bot==20.7
requests==2.31.0
python-dotenv==1.0.0
cryptography==41.0.7